Added error boundary for manga element
