+++ package.json
@@ -140,8 +140,10 @@
     "iso-639-1": "^2.1.15",
     "javascript-time-ago": "^2.5.9",
     "jotai": "^2.2.2",
+    "jotai-optics": "^0.3.0",
     "kute.js": "^2.2.4",
     "minidenticons": "^2.0.2",
+    "optics-ts": "^2.4.0",
     "prop-types": "^15.8.1",
     "protobufjs": "^7.2.4",
     "query-string": "^7.1.3",
