--- src/mangadex/pages/manga/Search/Content_Rating.tsx
@@ -1,1 +1,0 @@
-export type CttRtg_include = {name : string, include : boolean}
