--- src/mangadex/pages/manga/Search/Status_publ.tsx
@@ -1,2 +1,0 @@
-
-export type Status_include = {name : string, include : boolean}
