+++ src/mangadex/resources/componnents/mangas/MangaSearch/ReallySimpleCard.tsx
@@ -1,0 +1,12 @@
+import { Card, CardBody } from "@chakra-ui/react";
+import React from "react";
+
+export function ReallySimpleCard({ children }: React.PropsWithChildren) {
+    return (
+        <Card>
+            <CardBody>
+                {children}
+            </CardBody>
+        </Card>
+    );
+}
