+++ src/mangadex/resources/componnents/mangas/v1/MangaElementDef2_withChildren/vanilla.tsx
@@ -3,10 +3,10 @@ import React from "react";
 import { Link } from "react-router-dom";
 import TryCatch from "@commons-res/components/TryCatch";
 import { Manga } from "@mangadex/api/structures/Manga";
-import CoverImageByCoverID from "../../covers/v1/CoverImageByCoverID";
-import MangaTitle from "./MangaTitle";
+import CoverImageByCoverID from "../../../covers/v1/CoverImageByCoverID";
+import MangaTitle from "../MangaTitle";
 import { getMangaDexPath } from "@mangadex/index";
-import MangaContextMenu from "./MangaContextMenu";
+import MangaContextMenu from "../MangaContextMenu";
 
 const MangaDexPath = getMangaDexPath();
